+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+## [1.2.0] - 2025-11-12
+
 ### Changed
 
 - Global metrics prefix from `orderflow_proxy` to `flowproxy_`.
@@ -14,4 +16,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - Increase default order cache capacity to 1,048,576 entries (32 MiB).
 - Add `--io-threads` (`IO_THREADS`) and `--compute-threads` (`COMPUTE_THREADS`) CLI flags, defaults to 4 and 2 respectively.
 - Override dashes to underscore in value provided to CLI flag `--builder-name`.
+- Switch to HTTP/2 for system API (incoming and outgoing connections):
+    - Add `--http.client-pool-size` (`CLIENT_POOL_SIZE`) CLI flag, defaults to 8. HTTP/2 streams will be multiplexed over these clients.